Move data to separate folder
